blockchain: Move upgrade-only code to upgrade.go.
This moves the ticketsSpentInBlock and ticketsRevokedInBlock functions
to the upgrade.go file since they are only used in the upgrade code.
